requirements.txt
python-telegram-bot==20.6
openai==1.10.0
python-dotenv==1.0.0
APScheduler==3.10.4
nest_asyncio==1.5.8
